# service/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenID Federation Registry API",
    "description": "REST API for managing OpenID Federation entities, policies, modules, and trust marks. This API provides endpoints for creating, reading, updating, and deleting federation entities, policies, trust anchors, resolvers, trust marks, and trust mark subjects.",
    "contact": {
      "name": "Sweden Connect",
      "url": "https://www.swedenconnect.se",
      "email": "support@swedenconnect.se"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.5.10-SNAPSHOT"
  },
  "servers": [
    {
      "url": "https://registry.local.swedenconnect.se",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/registry/v1/{tenant}/{orgNumber}/entities/hosted/{entityId}": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "Get hosted entity",
        "operationId": "getHostedEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HostedEntity"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entities"
        ],
        "summary": "Update hosted entity",
        "operationId": "updateHostedEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostedEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HostedEntity"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Create hosted entity with specified ID",
        "operationId": "createHostedEntityWithId",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostedEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HostedEntity"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities"
        ],
        "summary": "Delete hosted entity",
        "operationId": "deleteHostedEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/registry/v1/{tenant}/{orgNumber}/entities/federation/{entityId}": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "Get federation entity",
        "operationId": "getFederationEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "includemodules",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FederationEntityWithModules"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entities"
        ],
        "summary": "Update federation entity",
        "operationId": "updateFederationEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FederationEntity"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Create federation entity with specified ID",
        "operationId": "createFederationEntityWithId",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FederationEntity"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities"
        ],
        "summary": "Delete federation entity",
        "operationId": "deleteFederationEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/registry/v1/{tenant}/{orgNumber}/entities/hosted": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "List hosted entity",
        "operationId": "listHostedEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "entityIdentifier",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostedEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Create hosted entity with auto-generated ID",
        "operationId": "createHostedEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostedEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HostedEntity"
                }
              }
            }
          }
        }
      }
    },
    "/registry/v1/{tenant}/{orgNumber}/entities/federation": {
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Create federation entity with auto-generated ID",
        "operationId": "createFederationEntity",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FederationEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/registry/v1/{tenant}/{orgNumber}/entities": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "List all entities",
        "description": "Lists all entities for the organization, optionally filtered by type and with modules included",
        "operationId": "listEntities",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includemodules",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithModules"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HostedEntity": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string",
            "format": "uuid",
            "description": "Entity ID",
            "readOnly": true
          },
          "entityIdentifier": {
            "type": "string",
            "description": "EntityIdentifier"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {},
            "description": "FederationMetadata"
          },
          "effectiveEcLocation": {
            "type": "string",
            "description": "Effective EcLocation that is calculated serverside",
            "readOnly": true
          },
          "trustMarkSources": {
            "type": "array",
            "description": "Trustmark sources that can be used to include trustmarks.",
            "items": {
              "$ref": "#/components/schemas/TrustmarkSource"
            }
          },
          "crit": {
            "type": "array",
            "description": "crit ",
            "example": "The crit (critical) Claim indicates that extensions to the set of Claims specified for use in this type of JWT",
            "items": {
              "type": "string"
            }
          },
          "authorityhints": {
            "type": "array",
            "description": "authorityhints",
            "example": "An array of strings representing the Entity Identifiers of Intermediate Entities or Trust Anchors that are Immediate Superiors of the Entity. the set of Claims specified for use in this type of JWT",
            "items": {
              "type": "string"
            }
          },
          "signingKeyId": {
            "type": "array",
            "description": "Key reference to signkeys from oidf-service",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "FederationEntityWithModules": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FederationEntity"
          },
          {
            "type": "object",
            "properties": {
              "entityId": {
                "type": "string",
                "format": "uuid",
                "description": "Entity ID",
                "readOnly": true
              },
              "entityIdentifier": {
                "type": "string",
                "description": "EntityIdentifier",
                "example": "https://ta.example.se/ta"
              },
              "crit": {
                "type": "array",
                "description": "crit ",
                "example": "The crit (critical) Claim indicates that extensions to the set of Claims specified for use in this type of JWT",
                "items": {
                  "type": "string"
                }
              },
              "authorityhints": {
                "type": "array",
                "description": "authorityhints",
                "example": "An array of strings representing the Entity Identifiers of Intermediate Entities or Trust Anchors that are Immediate Superiors of the Entity. the set of Claims specified for use in this type of JWT",
                "items": {
                  "type": "string"
                }
              },
              "signingKeyId": {
                "type": "array",
                "description": "Key reference to signkeys from oidf-service",
                "items": {
                  "type": "string"
                }
              },
              "trustAnchor": {
                "$ref": "#/components/schemas/TrustAnchor",
                "description": "Trust anchor module (if exists on federation entity)"
              },
              "intermediate": {
                "$ref": "#/components/schemas/Intermediate",
                "description": "Intermediate module (if exists on federation entity)"
              },
              "resolver": {
                "$ref": "#/components/schemas/Resolver",
                "description": "Resolver module (if exists on entity)"
              },
              "trustmarkIssuer": {
                "$ref": "#/components/schemas/TrustmarkIssuer",
                "description": "Trustmark issuer module (if exists on entity)"
              }
            }
          }
        ],
        "description": "Extended federation entity including modules"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token obtained from authentication endpoint",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}